requirements.txt
PyYAML~=6.0
requests~=2.28
urllib3~=1.26
charset-normalizer~=3.1
certifi~=2022.12
idna~=3.4
